Minimized public API
Pushing all business logic to the lower layers.
We should only change the public types when the public API changes.
